Add partial replacement of walkthrough
This is a lot at once, so many interconnected things… I’d
like to use provider components within games too, so I’ll
be migrating the capabilities walkthrough steps and adding
providers simultaneously.

There should be integration tests for the capabilities
walkthrough components so the details don’t need to be
in the acceptance test and it can be a mostly happy
path situation.

I had significant confusion where for a while I had
tasks in a capabilities service, but I ultimately
went back to provider components because it’s easier
to end things like decibel-monitoring when the provider
is destroyed. I still have confusion about the interaction
between tasks and single-result functions like
getCurrentPosition vs ongoing ones like DBMeter.

Other capabilities will work like DBMeter, where the
maximum is of interest, so there’ll probably be a
generalised component for that, and it can be used
in games too.